sharepoint/requirements.txt
requests
requests_ntlm
Office365-REST-Python-Client
pytest